Move 'preprocessOpenApiSpecs' call to docusaurus config
