+++ src/openapi/check-vendored-yaml.ts
@@ -1,3 +1,8 @@
+/*
+
+When making changes to this file, consider: https://virtru.atlassian.net/browse/DSPX-1577
+
+*/
 import * as fs from 'fs';
 import * as crypto from 'crypto';
 import { openApiSpecsArray } from './preprocessing';
