+++ src/openapi/update-vendored-yaml.ts
@@ -1,3 +1,8 @@
+/*
+
+When making changes to this file, consider: https://virtru.atlassian.net/browse/DSPX-1577
+
+*/
 import * as fs from 'fs';
 import * as path from 'path';
 import { openApiSpecsArray } from './preprocessing';
